chore: moves trigger methods to separate folder (#23574)
* chore: moves trigger methods to separate folder

* Changes

* Adds Partial typings to getTriggerChild invocation on popover

* Updates react-menu API

* Update change/@fluentui-react-menu-30e18336-5048-4a09-ac73-8df822fa4336.json
