+++ change/@fluentui-react-menu-30e18336-5048-4a09-ac73-8df822fa4336.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "fix: improve internal typings",
+  "packageName": "@fluentui/react-menu",
+  "email": "bernardo.sunderhus@gmail.com",
+  "dependentChangeType": "patch"
+}
