+++ change/@fluentui-react-popover-52fedd94-bdb6-4848-a41f-a3b8555f4922.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "Updates getTriggerChild internal usage",
+  "packageName": "@fluentui/react-popover",
+  "email": "bernardo.sunderhus@gmail.com",
+  "dependentChangeType": "patch"
+}
