+++ change/@fluentui-react-utilities-fe5f9201-bebc-4529-b6fb-9bad3c0779ac.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "chore: moves trigger methods to separate folder",
+  "packageName": "@fluentui/react-utilities",
+  "email": "bernardo.sunderhus@gmail.com",
+  "dependentChangeType": "patch"
+}
