+++ packages/react-components/react-popover/src/components/PopoverTrigger/usePopoverTrigger.ts
@@ -21,7 +21,9 @@ import type { PopoverTriggerChildProps, PopoverTriggerProps, PopoverTriggerState
  */
 export const usePopoverTrigger_unstable = (props: PopoverTriggerProps): PopoverTriggerState => {
   const { children } = props;
-  const child = React.isValidElement(children) ? getTriggerChild(children) : undefined;
+  const child = React.isValidElement(children)
+    ? getTriggerChild<Partial<PopoverTriggerChildProps>>(children)
+    : undefined;
 
   const open = usePopoverContext_unstable(context => context.open);
   const setOpen = usePopoverContext_unstable(context => context.setOpen);
