+++ packages/react-components/react-utilities/src/utils/index.ts
@@ -1,9 +1,6 @@
-export * from './applyTriggerPropsToChildren';
 export * from './clamp';
 export * from './getNativeElementProps';
 export * from './getRTLSafeKey';
-export * from './getTriggerChild';
-export * from './isFluentTrigger';
 export * from './omit';
 export * from './properties';
 export * from './shouldPreventDefaultOnKeyDown';
